requirements.txt
discord.py
pyyaml
python-dateutil
aiohttp